Reworked architecture of scripts

+++ UninstallLocal.ps1
@@ -1,34 +1,35 @@
 param(
-    [Parameter()] [string] $pkgname,
-    [Parameter()] [string] $archive_path
+    [Parameter()] [string] $pkgname
 );
 
 $ErrorActionPreference = "Stop";
 
-$base_dir = $PSScriptRoot;
-$ports_dir = "$base_dir\ports";
-$assets_file = "$base_dir\local_assets.json";
-$functions_file = "$base_dir\Functions.ps1";
+# Source functions & variables
+. $PSScriptRoot\Functions.ps1;
 
-# Source functions
-. $functions_file;
-
-# Retrieve local assets list
-$local_assets = Get-Content -Path $assets_file | ConvertFrom-Json;
-if (!($local_assets -is [array])) {
-    $local_assets = ,$local_assets;
+# Check if pkg is installed
+$to_remove = @();
+$installed = Pkg-List $pkgname;
+if (!$installed) {
+    Write-Warning "'$pkgname' is not installed."
+    return;
 }
-# Check if pkg is listed
-$pkg_is_listed = !(!($local_assets | ?{$_ -eq $pkgname}));
 
-if (!$pkg_is_listed) {
-    Write-Error "'$pkgname' is locally installed. It's possible a classic install overwrote the local one."
+# Retrieve local versions
+$installed | %{
+    $version = ((vcpkg list $_).Split(" ") | ?{$_})[1];
+    if ($version -match "local_") {
+        $to_remove += $_;
+    }
+}
+
+if (!$to_remove) {
+    Write-Warning "'$pkgname' is installed, but as a release version and not locally."
+    return;
 }
 
 # Remove pkg
-Pkg-Remove $pkgname;
+vcpkg remove $to_remove;
+
 # Remove pkg folder
 Remove-Item -Recurse -Force "$ports_dir\$pkgname";
-# Unlist pkg from json
-$local_assets = $local_assets | ?{$_ -ne $pkgname};
-$local_assets | ConvertTo-Json | Out-File $assets_file;
